backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
PyJWT
pwdlib
pydantic
pydantic-settings
alembic
pytest
httpx